+++ lib/formotion/row_type/text_row.rb
@@ -20,6 +20,7 @@ module Formotion
         field.autocapitalizationType = row.auto_capitalization if row.auto_capitalization
         field.autocorrectionType = row.auto_correction if row.auto_correction
         field.placeholder = row.placeholder
+        field.enabled = row.editable?
 
         field.on_begin do |text_field|
           row.on_begin_callback && row.on_begin_callback.call
@@ -72,6 +73,9 @@ module Formotion
       end
 
       def on_select(tableView, tableViewDelegate)
+        if !row.editable?
+          return
+        end
         field.becomeFirstResponder
       end